backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.9
opencv-python-headless==4.8.0.74
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
